corrected functiondeclaration, Makefile

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: lmeubrin <lmeubrin@student.42berlin.       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/22 15:20:25 by lmeubrin          #+#    #+#             */
-/*   Updated: 2024/04/23 16:53:59 by lmeubrin         ###   ########.fr       */
+/*   Updated: 2024/04/25 11:10:35 by lmeubrin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ size_t	ft_strlen(const char *s);
 
 size_t	ft_strlcpy(char *dest, const char *src, size_t size);
 
-size_t	ft_strlcat(char *dest, const char *src, unsigned int size)
+size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
 	return (ft_strlcpy((dest + ft_strlen(dest)), src, size));
 }
